server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiomysql==0.2.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
